+++ src/main/java/org/folio/InventoryKafkaTopic.java
@@ -10,8 +10,14 @@ public enum InventoryKafkaTopic implements KafkaTopic {
   INSTANCE_CONTRIBUTION("instance-contribution"),
   BOUND_WITH("bound-with"),
   ASYNC_MIGRATION("async-migration"),
-  SERVICE_POINT("service-point");
+  SERVICE_POINT("service-point"),
+  CLASSIFICATION_TYPE("classification-type");
 
+  private static final String DEFAULT_NUM_PARTITIONS_PROPERTY = "KAFKA_DOMAIN_TOPIC_NUM_PARTITIONS";
+  private static final String DEFAULT_NUM_PARTITIONS_VALUE = "50";
+  private static final String CLASSIFICATION_TYPE_NUM_PARTITIONS_PROPERTY =
+    "KAFKA_CLASSIFICATION_TYPE_TOPIC_NUM_PARTITIONS";
+  private static final String CLASSIFICATION_TYPE_NUM_PARTITIONS_VALUE = "1";
   private final String topic;
 
   InventoryKafkaTopic(String topic) {
@@ -30,9 +36,18 @@ public enum InventoryKafkaTopic implements KafkaTopic {
 
   @Override
   public int numPartitions() {
-    return Integer.parseInt(StringUtils.firstNonBlank(
-      System.getenv("KAFKA_DOMAIN_TOPIC_NUM_PARTITIONS"),
-      System.getProperty("KAFKA_DOMAIN_TOPIC_NUM_PARTITIONS"),
-      System.getProperty("kafka-domain-topic-num-partitions"), "50"));
+    if (this == CLASSIFICATION_TYPE) {
+      return getNumberOfPartitions(CLASSIFICATION_TYPE_NUM_PARTITIONS_PROPERTY,
+        CLASSIFICATION_TYPE_NUM_PARTITIONS_VALUE);
+    }
+    return getNumberOfPartitions(DEFAULT_NUM_PARTITIONS_PROPERTY, DEFAULT_NUM_PARTITIONS_VALUE);
   }
+
+  private int getNumberOfPartitions(String propertyName, String defaultNumPartitions) {
+    return Integer.parseInt(StringUtils.firstNonBlank(
+      System.getenv(propertyName),
+      System.getProperty(propertyName),
+      System.getProperty(propertyName.toLowerCase().replace('_', '-')), defaultNumPartitions));
+  }
+
 }